version 1.1
* some help files
  click Help button on right end of menu button row
  incomplete

* read air tile bitmaps into memory using a background thread
  it would take too much gui thread time when zoomed out

* fix some compiler warnings (unused result, empty catch blocks, redundant casts, etc)

* make SQLite query() calls consistent format
  and put column name arrays in static constants

* Maint/Plate screen can be formatted as a List or Map using an Option

* preload OpenStreetMap tiles during database download so they will be
  available without an Internet connection (provided the preload has
  had time to complete)

* fill in the plate unload code so per-state data can be unloaded
  if it is no longer wanted

* make sure GPS NOT UPDATING message fits the screen in case screen
  is small

+++ app/src/main/java/com/outerworldapps/wairtonow/GlassView.java
@@ -339,7 +339,7 @@ public class GlassView
 
         float spdflt = currPos.speed * KtPerMPS;
         if (wairToNow.optionsView.ktsMphOption.getAlt ()) spdflt *= SMPerNM;
-        int spdint = (int) Math.round (spdflt);
+        int spdint = Math.round (spdflt);
         DrawNumericStrip (canvas, aspCentX, aspCentY, aspHeight, aspWidth, spdint, 10, 8, 0, -1, -1);
 
         int feet = (int) Math.round (altitude * 3.28084);
